pci: layerscape: Fix a compilation warning on ls102xa
The fdt_size_t has different byte width on arm64 and arm32 platforms.
Fix the following warning by converting the type of 'cfg_size' to
'u64' forcedly.
drivers/pci/pcie_layerscape.c: In function ‘ls_pcie_probe’:
drivers/pci/pcie_layerscape.c:573:40: warning: format ‘%llx’ expects argument of type ‘long long unsigned int’, but argument 4 has type ‘fdt_size_t’ {aka ‘long unsigned int’} [-Wformat=]
  573 |   printf("PCIe%d: %s Invalid size(0x%llx) for resource \"config\",expected minimum 0x%x\n",
      |                                     ~~~^
      |                                        |
      |                                        long long unsigned int
      |                                     %lx
  574 |          PCIE_SRDS_PRTCL(pcie->idx), dev->name, cfg_size, SZ_8K);
      |                                                 ~~~~~~~~
      |                                                 |
      |                                                 fdt_size_t {aka long unsigned int}

Fixes: c481b5c86b2c ("pci: layerscape: Add size check for config resource")

+++ drivers/pci/pcie_layerscape.c
@@ -571,7 +571,8 @@ static int ls_pcie_probe(struct udevice *dev)
 	cfg_size = fdt_resource_size(&pcie->cfg_res);
 	if (cfg_size < SZ_8K) {
 		printf("PCIe%d: %s Invalid size(0x%llx) for resource \"config\",expected minimum 0x%x\n",
-		       PCIE_SRDS_PRTCL(pcie->idx), dev->name, cfg_size, SZ_8K);
+		       PCIE_SRDS_PRTCL(pcie->idx), dev->name,
+		       (u64)cfg_size, SZ_8K);
 		return 0;
 	}
 	/*
